www/backends/shares/requirements/base.txt
pyjwt[crypto]
itsdangerous
mjml
aiosmtplib
amqp
annotated-types
argon2-cffi
asgiref
asyncpg
babel==2.16.0
bcrypt==4.0.1
celery
cloudinary
dnspython
email-validator
fastapi
fastapi-cli
fastapi-mail
hiredis
humanize
jinja2
loguru
mjml-python
passlib
paystackapi==2.1.3
phonenumbers==8.13.47
pillow
premailer
pydantic
pydantic-settings
pydantic-extra-types==2.9.0
python-dateutil
python-jose
python-magic
python-multipart
python-slugify
redis
requests
resend
sqlalchemy
sqlmodel
typing-extensions
uvicorn
uvicorn-worker
urllib3
websockets
yfinance==0.2.44
